requirements.txt
zxcvbn==4.4.28
requests==2.31.0
fpdf2==2.7.4